Moved Json,TaskExtensions and Singleton to core project
