+++ templates/Uwp/Features/BackgroundTask.Prism/Param_ProjectName/BackgroundTasks/BackgroundTask.cs
