+++ templates/Uwp/Features/BackgroundTask.Prism/Param_ProjectName/BackgroundTasks/BackgroundTaskFeature.cs
