+++ templates/Uwp/Features/BackgroundTask.Prism/Param_ProjectName/Services/BackgroundTaskService$BackgroundTaskFeature_gpostaction.cs
