+++ templates/Uwp/Features/BackgroundTask.Prism/Param_ProjectName/Services/IBackgroundTaskService.cs
