+++ templates/Uwp/Features/BackgroundTask/Param_ProjectName/BackgroundTasks/BackgroundTask.cs
