+++ templates/Uwp/Features/BackgroundTask/Param_ProjectName/BackgroundTasks/BackgroundTaskFeature.cs
