+++ templates/Uwp/Features/BackgroundTask/Param_ProjectName/Services/ActivationService$CleanUpInitializeAsync_gpostaction.cs
