+++ templates/Uwp/Features/BackgroundTaskVB/Param_ProjectName/Services/ActivationService$CleanUpInitializeAsync_gpostaction.vb
