+++ templates/Uwp/Features/LiveTile/Services/ActivationService_postaction.cs
@@ -6,7 +6,7 @@
 
 using System;
 //{[{
-using Param_RootNamespace.Helpers;
+using Param_RootNamespace.Core.Helpers;
 //}]}
 
 namespace Param_ItemNamespace.Services
