+++ templates/Uwp/Features/LiveTileVB/Services/ActivationService_postaction.vb
@@ -5,7 +5,7 @@
 '**}
 
 '{[{
-Imports Param_RootNamespace.Helpers
+Imports Param_RootNamespace.Core.Helpers
 '}]}
 
 Namespace Services
