+++ templates/Uwp/Features/SettingsStorageVB/Param_ProjectName/Helpers/SettingsStorageExtensions.vb
@@ -1,5 +1,6 @@
 ﻿Imports Windows.Storage
 Imports Windows.Storage.Streams
+Imports Param_ItemNamespace.Core.Helpers
 
 Namespace Helpers
     ' Use these extension methods to store and retrieve local and roaming app data
