+++ templates/Uwp/Features/ToastNotifications/Activation/DefaultLaunchActivationHandler_postaction.cs
@@ -2,7 +2,7 @@
 //This code block adds code to show a sample toast notification on application start to your project.
 //**}
 //{[{
-using Param_RootNamespace.Helpers;
+using Param_RootNamespace.Core.Helpers;
 using Param_RootNamespace.Services;
 //}]}
