+++ templates/Uwp/Projects/Default/wts.DefaultProject.csproj
@@ -131,7 +131,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Helpers\ResourceExtensions.cs" />
-    <Compile Include="Helpers\Singleton.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Models\" />
