+++ templates/Uwp/_composition/_shared/Feature.DeepLinking.AddActivationHandlerVB/Services/ActivationService_postaction.vb
@@ -1,5 +1,5 @@
 ﻿'{[{
-Imports Param_RootNamespace.Helpers
+Imports Param_RootNamespace.Core.Helpers
 '}]}
 Namespace Services
     Friend Class ActivationService
