+++ templates/Uwp/_composition/_shared/Project.AddCoreProject/Param_ProjectName.Core/Helpers/Singleton.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Collections.Concurrent;
 
-namespace wts.DefaultProject.Helpers
+namespace Param_RootNamespace.Core.Helpers
 {
-    internal static class Singleton<T>
+    public static class Singleton<T>
         where T : new()
     {
         private static ConcurrentDictionary<Type, T> _instances = new ConcurrentDictionary<Type, T>();
